// src/app/models/app.ts
export class App {
  build_status: string;
  app_name: string;
  deploy_version: string;
  monitor_status: string;
  deploy_time: string;
  app_id: string;
  build_time: string;
  build_errors_count: number;
  build_version: string;
  build_build_number: number;
  deploy_build_number: number;
  build_id: string;
  deploy_warnings_count: number;
  deploy_count: number;
  env: string;
  warnings_count: number;
  build_count: number;
  monitor_errors_count: number;
  monitor_warnings_count: number;
  errors_count: number;
  deploy_id: string;
  build_warnings_count: number;
  deploy_errors_count: number;
  monitor_deploy_count: number;
  deploy_status: string;
  build_commit_count: number;
  build_days: string;
}